Moderate styling additions
Added classes to the output of usernames/passwords etc so that they
showed a little more visually what was accepted or in/correct etc.
Removed spellcheck from the texture (know it might be removed but just
so unknown words don’t get underlined by red in the demo). Minor
spacing adjustments to table/text output.